backend/requirements.txt
fastapi
uvicorn[standard]

httpx

sqlalchemy
psycopg2-binary
alembic

redis

pydantic
pydantic-settings

python-dotenv

orjson